Deprecate external image related configuration in ParserOptions
Per-parser configuration is discouraged; use sitewide configuration instead.

Code search:
https://codesearch.wmflabs.org/search/?q=setEnableImageWhitelist%7CsetAllowExternalImages%7CsetAllowExternalImagesFrom&i=nope&files=&repos=

The ultimate goal here is to refactor the image filtering functionality into
an extension and move it out of core, so that it can be used by both
Parsoid and the legacy parser in the same way.  We may add back per-parser
customization of the filtering, but the API will probably look different.
Deprecate the existing ParserOptions-based mechanism, which code search
indicates almost no one is using.

Bug: T254802
Change-Id: Ib4a59bbae10cfc924c0290948330d93e02de9ed0