maintenance/blockUsers.php: Check only for prior DatabaseBlocks
BlockUsers needs to know if there is a block stored in the database
against a particular user account. This is answered by
DatabaseBlock::newFromTarget.

Prior to this, the script called User::getBlock, which can return
SystemBlocks and CompositeBlocks too. These are not stored in the
database, so are irrelevant and would cause an error on attempting to
update them.

It also meant parameter types for BlockIpCompleteHook needed to be
changed to AbstractBlock in a302833fc35ad81e4cc854404ef00835da1b2b3a.
This changes them back to DatabaseBlock.

Change-Id: Iec90d4938e76d76ac1f5b1d1e9ffb736a58d5c53