checkboxHack: Trigger 'input' event when checkbox state changes
This event is necessary to trigger saving the checkbox state, but
programmatic changes to checkbox.checked do not trigger an input or change event.

function setCheckedState( checkbox, checked ):
Sets the checked state and fires the 'input' event.
The event bubbles, consistent with the standard behavior of the input event.

setCheckedState() is called when a user event on
some element other than the checkbox
should result in changing the checkbox state.

Bug: T246427
Bug: T255727
Change-Id: Ib8e88dfd9d3a1431ae73b8fed0f540f89834ae7f