PLATFORM-4871 | Add ParserAllowExternalImage hook in Parser::maybeMakeExternalImage method
Currently `Parser` allows only hot linking images with ``.jpg|.jpeg|.png|.gif` in the end of url. As we use revision, version and query params in our Vignette urls, so they would not be allowed for render even if domain was added to `wgAllowExternalImagesFrom`. `ParserAllowExternalImage` hook will be used only if domain in image url is added to `wgAllowExternalImagesFrom` list or `wgAllowExternalImages` is set to `true`.