Removed findExtensions from Installer
There was an override for findExtensions which took care of the nested BlueSpice 2 extension folder structure (we had all extensions in     BlueSpiceExtensions/EXTENSIONNAME back then). This is no longer the case and the override is no longer needed. It also currently returns an array instead of a status object. So for easier future compatibility, we remove the override.
    
    [3.2]
    ERM#19790