findBadBlobs: better separate scan and mark modes.
This makes the following changes to the findBadBlobs utility:
- rename --from-date to --scan-from, to match the intended use.
- require the usage of --revisions with --mark, so revisions
  cannot be marked directly when found by a scan.
- catch any exception when testing for bad blobs, casting
  a wider net.
- change the output format, so the IDs of bad revisions can easily be
  extracted by command line tools for further processing.
- warn when trying to mark blobs that can successfully we read.

The idea is to allow detection of blobs that are "bad" in a
large variety of ways, including due to misconfiguration, while at the
same time making sure that blobs do not get marked as bad due to
temporary outages.

The intended usage of findBadBlobs is to first scan a potentially
problematic set of revisions using --scan-from, review to errors found,
and then determine which of the revisions should be marked as bad.
Once the bad revisions have been identified, a list with their IDs
can be extracted from the output, and supplied back to findBadBlobs
via the --revisions option.

Bug: T251778
Change-Id: I47c11190b665c1dac88db32ee2bf683728cb3dc6