Fix error handling in TextSlotDiffRenderer::getTextDiffInternal()
A long time ago, returning false from the predecessor of this
method (done on filesystem error when trying to generate a diff
via an external executable) meant the diff result was treated as
an internal error and not cached, but I2f8a9db broke that.

There is not much point in trying to fail softly on tempfile
creation errors anyway, so just replace it with exception throwing.

Change-Id: Idb15a0e1f67cb1ccbf7c47788f4eecbaa8176665