Deprecate $wgUseTwoButtonsSearchForm
This variable is not used in core, but it is in at least 7 skins
including 3 deployed in Wikimedia production.

And since its on by default MediaWiki setting, probably a large
number of users are accustomed to what it provides, especially
given the fact that there's no way they could have turned it off.

So it will be deprecated now and when it is removed completely,
skins that want continue supporting it then have to define it as
their own config variable. Discussion on this tracked in T256568

Bug: T42622
Change-Id: I04c7627420bfa1b1b80ada3a561f8fcd159fbdf7