concatenateChunks failures are not job failures
concatenateChunks can fail for many reasons, including
- corrupt files
- don't pass abusefilter
- exceeds file limits
etc.

When we're unable to concatenate the chunks, it shouldn't be
considered a job failure. It just didn't pass whatever requirements,
and the user should be informed about it, but the job actually
did what it had to do and it's not the software's fault that the
file isn't any good.

Besides, this job is only triggered on async uploads.
Synchronous uploads call concatenateChunks directly, and no exceptions
are thrown there when it fails to concatenate: we simply inform the
users what happened - same as what's already happening for these async
uploads as well.

TL;DR: The job did what it has to do.

Bug: T204881
Change-Id: Idf166d2fb509211080fb9e5f0e085467ef7d7ef2