OutputPage: Use @unstable instead of @internal for addHtmlClasses()
Follows-up 9d3254d18e7e.

Methods using @internal are designed for use internal to the component
that defines the method, and are generally considered safe to remove
anytime if they are unused.

This method however is not for internal use, it is designed for public
use (nothing in core needs it), but we're not yet ready to commit
to it being a stable API that would receive no breaking changes or
be removed without deprecation.

The appropiate tag for such experimental feature that allows early
adoption (with caution, knowing it might get renamed, moved or dropped
in the next major release) – is @unstable.
<https://www.mediawiki.org/wiki/Stable_interface_policy>

Bug: T247862
Change-Id: Iceaaf520979afa1c3ee806262083f94d67660174