Handle undos in action=mcrundo properly
This aims to fix the following:
* Properly handle the undoafter param (it is currently set as the
oldest reverted revision ID, which is BAD)
* Set $originalRevId on undos to let EditResultBuilder check
whether the undo was an exact one and to let extensions know
about the "base" revision for this undo.

Note: $originalRevId was previously set explicitly set to false
here which does not make any sense to me. Setting it to the
revision we base the undo on is the only way to make EditResult's
getters' values consistent and make any sense. According to
PageUpdater's docs, this should only be set when the new revision has
"the exact same content as the given original revision", which is
checked here. I don't necessarily agree with this view, but it's
the spec, so we should probably keep to it.

I also hope it to be a kind of a "reference" implementation for
undos in case someone tries to tidy up the EditPage class in the
future, so I left quite a few comments behind.

I wrote some integration tests for that, should be sufficient for
now. I think.

Bug: T256915
Change-Id: I1c1f8d3b8a6273fbe7b460d140ef468f560456ce