rdbms: Mark the Database class as stable for subclassing
Strategy used to determine which methods should be stable:
- Methods that encode specific SQL syntax
- Methods already overwritten by a subclass

Strategy used to determine which methods should NOT be stable:
- private or final (obviously)
- encodes logic for managing database connections
- wraps a call to a method that is intended for overriding

Note that such methods would be overwritten by an extension that
provides support for a specific database backend. Such extensions
are only now becoming possible in 1.35. Marking methods in Database
as stable for overriding is done to support the creation of
such extension (e.g. for MSSQL and Oracle).

Bug: T247862
Change-Id: Ief5386c0ff4239a8d95db2d6896338612aa56bb0