Use OOUI to toggle watchthis checkbox
Use OOUI instead of jQuery to toggle the edit form's watch-this
checkbox when the page's watch status is changed via the watchstar
link.

Because of this OOUI events will be fired, and the expiry dropdown
will correctly disable when the watch-this checkbox is toggled.

This is a follow-up to 4fa7d4d7a707e9bcd6646d5272a4b0c2842d6ea6

Bug: T248494
Change-Id: I6aa11d703a08f2f9b0dbbe29b0145ced3c5167b8