Set $wgShellRestrictionMethod to 'autodetect' by default
In practice shell restrictions are still opt-in because most people
won't have firejail installed. But it makes it more straightforward for
those who do want the extra security and it makes it easier for
distributors to provide a more secure environment by default.

Also include the release notes for 24ddc62a3c9ec2c.

Change-Id: Idc08d2d1d17d6ea2d9e4dd0cc7bd85749308c195