shell: Make ->restrict( RESTRICT_NONE ) actually work
Have ->restrict() overwrite any previous restrictions instead of adding
to the existing list. Multiple examples are provided on how this
function should be called going forward.

According to codesearch, all non-test uses of ->restrict() were already
expecting this behavior, passing values like:
 Shell::RESTRICT_DEFAULT | Shell::NO_NETWORK
when trying to disable network access.

This is a breaking change, but IMO one that is going to fix more things
than it breaks.

Bug: T257278
Change-Id: I1895d1fc73cc793af2f82001e9d5874b7520f802