Deprecate printableversion=yes
Replace the link to javascript:print() and add a warning message
on the printableversion when used.

The link is hidden if JS is not enabled and a skin includes the
'interface' ResourceLoaderSkinModule feature.

As part of this change we also need to notify users that
printableversion is no longer supported. To do this we need
to put a warningbox at the top of the page.

While this should be straightforward, as warning b ox styles
are loaded by all skins via the legacy module, it does require a minor
change to make those CSS rules apply inside print media as they are
currently scoped to screen. We rectify this, by splitting the rules
out from mediawiki.legacy.skinning and applying them to all media
queries. This means warning boxes will appear styled in print media
however these are seldom used in article namespaces - and when they are
they are visible - it would be better they look like warning boxes!

Bug: T167956
Change-Id: If1a3cfc6d82e9e389be7bf17fe288e212aa64139
(cherry picked from commit 06cb860a727d43500ee92a7bf055cdf757364ff8)