preferences: Prefer new kind of warnings about invalid HTML
In df7231ad8915922e85437fbcdee114eed8c987d9 we added a new system for
validating signatures, but there is already an old one that can catch
some simple mistakes.

The new system prints more details about the error and how to correct
it, so if both systems report an error, display the one from the new
system.

Bug: T140606
Change-Id: I731429a56ba191c25b5b344a9dc9282106e5e2a6