Return directly from within try/finally block
It is good code style to return a variable from the same scope as it is
definied and not from a level deeper.
For try/finally blocks this does not change the behaviour because the
finally block is executed if the scope is leaved, normally or by return
or by thrown exception.

This helps static code analyzer like phan to understand the code better.

Change-Id: Ib947e58adb3676b1d69507a15ea44d93c5a9973d