Default to using watchlist expiry of old page when moving pages
Expiries are copied post-send so as to not slow anything down.

Note the expiry feature has been disabled in some unit tests that mock
database queries, simply because these tests are so hard to maintain and
are very fragile. The logic introduced with this patch is covered by the
integration tests.

Bug: T257259
Change-Id: I4223eaa6782a319fb684acf656f54b88a92e5288