Watchlist: Fix action=watch caching the wrong expiry
WatchedItem was being cached with an invalid expiry always
resulting in isExpired()* returning true and a query to the
replica db being executed also returning the wrong expiry due to
replication lag in the same request since the record is being
updated a little bit earlier in the call-stack.

This was not a problem when `$wgWatchlistExpiry` was disabled
because isExpired()* is never true and the cache object was
always returned avoiding the query to the replica.

* See ::getWatchedItem() which gets called indirectly from
  SkinTemplate after the update to the db has been made on
  addWatchBatchForUser

Bug: T253937
Change-Id: Ic4d186433c4fc105a3268d6cf98b8ef8ce023cd4