mw.widgets.UsersMultiSelectWidget: Don't assume `suggestions` array element exists
I45632416 made the UsersMultiSelectWidget more accessible by making it
signal its active child to assistive technologies when the list of
suggested usernames is updated.

However, the change introduced a bug where an element in the list of
suggested usernames is accessed even when the list is empty, which
triggers an error.

Bug: T250762
Change-Id: I517dd0669bfe6137a2b601addf288dcd5135c809