Don't output HTMLDebugLog div element when not needed
Add div wrapper to MWDebug::getHTMLDebugLog() to allow users to fully
style the whole block, in addition to the id to style the list items.

It's moved to MWDebug to ensure its present in all skins instead of
using duplication for both Mustache-based and SkinTemplate-based Skins.

The wrapper is only added when $wgShowDebug is true, that's when the
function returns non-empty string

Bug: T259075
Change-Id: I0c387573bdd8ab72e4b2338b802df8b7881036da