Fix parserTests.php by ensuring change_tag tables are cloned
This fixes a regression in the parser test CLI runner caused by
435d5f4d55d3d88880c44978e7aeabe740ad5635, which added a change tag for
"manual reverts".  The parser test framework was triggering this
change tag addition as it set up its mock article database, and then
subsequent attempts to write the change tag to the database failed
because the tables were missing.

Bug: T259186
Change-Id: I232e918dfdc83244a010681b6adffd6c1171cf24