Add ban-bali to list of known language codes
This change adds ban-bali (Balinese language in Balinese script), falling back to ban (Balinese language in Latin script).

The language code ban-bali is already known to UniversalLanguageSelector and will be used on Balinese Wikisource (currently in development, hosted on Multilingual Wikisource). ban-bali will likely be used as a page language only for now, or for pieces of text within pages. It's important to make it available as a page language so that ULS can apply the correct font.

Bug: T245359
Change-Id: Icaff92904e9d1250c8c84bfcc6cfa3ebcb5230bd